refactor(chart): make ChartEx renderer opt-in
Keep classic DrawingML 2-D chart families in the default renderer while moving Microsoft ChartEx family painting behind a separate entry and injected renderer contract. Wire the contract through main and worker modes for DOCX, XLSX, and PPTX without changing parser WASM packaging.

Add bundle-boundary, renderer, worker, public API, documentation, and consumer integration coverage. The default main-mode format bundles no longer reach the ChartEx implementation, while the self-contained worker assets preserve existing bundler compatibility.
