+++ packages/core/src/worker/renderer-module-contract.ts
@@ -4,7 +4,7 @@
  * public renderer interfaces. */
 const WORKER_RENDERER_MODULE_PROTOCOL = 'ooxml-worker-renderer-module/v1' as const;
 
-export type WorkerBuiltinRendererName = 'math' | 'threeD' | 'regionMap';
+export type WorkerBuiltinRendererName = 'math' | 'threeD' | 'regionMap' | 'chartEx';
 
 interface WorkerBuiltinRendererDescriptorBase {
   readonly protocol: typeof WORKER_RENDERER_MODULE_PROTOCOL;
@@ -19,7 +19,7 @@ interface WorkerMathRendererDescriptor extends WorkerBuiltinRendererDescriptorBa
 }
 
 interface WorkerChartRendererDescriptor extends WorkerBuiltinRendererDescriptorBase {
-  readonly builtin: 'threeD' | 'regionMap';
+  readonly builtin: 'threeD' | 'regionMap' | 'chartEx';
 }
 
 export type WorkerRendererDescriptor =
@@ -30,12 +30,14 @@ export interface WorkerRendererDescriptors {
   readonly math?: WorkerRendererDescriptor;
   readonly threeD?: WorkerRendererDescriptor;
   readonly regionMap?: WorkerRendererDescriptor;
+  readonly chartEx?: WorkerRendererDescriptor;
 }
 
 export interface WorkerRendererSources {
   readonly math?: object;
   readonly threeD?: object;
   readonly regionMap?: object;
+  readonly chartEx?: object;
 }
 
 const workerRendererRegistry = new WeakMap<object, WorkerRendererDescriptor>();
@@ -46,7 +48,7 @@ function createBuiltinWorkerRendererDescriptor(
   engineAssetUrl: string,
 ): WorkerMathRendererDescriptor;
 function createBuiltinWorkerRendererDescriptor(
-  builtin: 'threeD' | 'regionMap',
+  builtin: 'threeD' | 'regionMap' | 'chartEx',
 ): WorkerChartRendererDescriptor;
 function createBuiltinWorkerRendererDescriptor(
   builtin: WorkerBuiltinRendererName,
@@ -70,7 +72,7 @@ export function registerBuiltinWorkerRenderer<T extends object>(
 ): T;
 export function registerBuiltinWorkerRenderer<T extends object>(
   renderer: T,
-  builtin: 'threeD' | 'regionMap',
+  builtin: 'threeD' | 'regionMap' | 'chartEx',
 ): T;
 export function registerBuiltinWorkerRenderer<T extends object>(
   renderer: T,
@@ -92,7 +94,8 @@ export function assertWorkerRendererDescriptor(
   }
   if (descriptor.builtin !== 'math'
     && descriptor.builtin !== 'threeD'
-    && descriptor.builtin !== 'regionMap') {
+    && descriptor.builtin !== 'regionMap'
+    && descriptor.builtin !== 'chartEx') {
     throw new TypeError(`Unsupported built-in worker renderer: ${String(descriptor.builtin)}`);
   }
   if (descriptor.builtin === 'math' && typeof descriptor.engineAssetUrl !== 'string') {
@@ -110,10 +113,12 @@ export function workerRendererDescriptors(
   const math = sources.math ? workerRendererRegistry.get(sources.math) : undefined;
   const threeD = sources.threeD ? workerRendererRegistry.get(sources.threeD) : undefined;
   const regionMap = sources.regionMap ? workerRendererRegistry.get(sources.regionMap) : undefined;
+  const chartEx = sources.chartEx ? workerRendererRegistry.get(sources.chartEx) : undefined;
   const descriptors: WorkerRendererDescriptors = {
     ...(math ? { math } : {}),
     ...(threeD ? { threeD } : {}),
     ...(regionMap ? { regionMap } : {}),
+    ...(chartEx ? { chartEx } : {}),
   };
   return Object.keys(descriptors).length > 0 ? Object.freeze(descriptors) : undefined;
 }
