+++ packages/core/src/worker/renderer-module.test.ts
@@ -60,6 +60,7 @@ describe('worker renderer module descriptors', () => {
       }),
       threeD: registerBuiltinWorkerRenderer({}, 'threeD'),
       regionMap: registerBuiltinWorkerRenderer({}, 'regionMap'),
+      chartEx: registerBuiltinWorkerRenderer({}, 'chartEx'),
     };
 
     const loaded = await loadWorkerRenderers(workerRendererDescriptors(sources));
@@ -67,6 +68,7 @@ describe('worker renderer module descriptors', () => {
     expect(typeof loaded.math?.mathMLToSvg).toBe('function');
     expect(typeof loaded.threeD?.render).toBe('function');
     expect(typeof loaded.regionMap?.render).toBe('function');
+    expect(typeof loaded.chartEx?.render).toBe('function');
   });
 
   it('carries the consumer-resolved math asset URL without exposing it on the renderer', () => {
