+++ packages/pptx/api/public-api-baseline.d.ts
@@ -579,6 +579,9 @@ export interface ChartRect {
     w: number;
     h: number;
 }
+export interface ChartExRenderer {
+    render(ctx: CanvasRenderingContext2D, chart: ChartModel, rect: ChartRect, ptToPx: number, shapeRotationDeg?: number): boolean;
+}
 export interface ChartRegionMapRenderer {
     render(ctx: CanvasRenderingContext2D, chart: ChartModel, rect: ChartRect, ptToPx: number, shapeRotationDeg?: number): boolean;
 }
@@ -988,6 +991,7 @@ interface LoadOptions__emitterCollision1 {
     math?: MathRenderer;
     threeD?: ChartThreeDRenderer;
     regionMap?: ChartRegionMapRenderer;
+    chartEx?: ChartExRenderer;
 }
 export interface MatchRunSlice {
     runIndex: number;
@@ -1696,6 +1700,7 @@ export type SlideRenderOptions = RenderOptions & {
     math?: MathRenderer;
     threeD?: ChartThreeDRenderer;
     regionMap?: ChartRegionMapRenderer;
+    chartEx?: ChartExRenderer;
     dim?: DimOptions;
 };
 export interface SoftEdge {
