+++ packages/pptx/src/index.ts
@@ -60,6 +60,7 @@ export type {
   ChartThreeDSeriesAxis,
   ChartThreeDRenderer,
   ChartRegionMapRenderer,
+  ChartExRenderer,
   ChartExElementStyle,
   ChartLineDashSegment,
   ChartexHistogramBinning,
