+++ packages/pptx/src/scroll-viewer.ts
@@ -521,6 +521,7 @@ export class PptxScrollViewer implements ZoomableViewer {
         math: this._opts.math,
         threeD: this._opts.threeD,
         regionMap: this._opts.regionMap,
+        chartEx: this._opts.chartEx,
         mode: this._mode,
       }), (ownedPresentation) => {
         // Invalidate before TerminalResourceOwner installs the candidate and
