+++ packages/xlsx/src/index.ts
@@ -61,6 +61,7 @@ export type {
   ChartThreeDSeriesAxis,
   ChartThreeDRenderer,
   ChartRegionMapRenderer,
+  ChartExRenderer,
   ChartExElementStyle,
   ChartLineDashSegment,
   ChartexHistogramBinning,
