+++ packages/xlsx/src/worker-render-deps.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 import type {
   ChartRegionMapRenderer,
   ChartThreeDRenderer,
+  ChartExRenderer,
   MathRenderer,
 } from '@silurus/ooxml-core';
 import type { ParsedWorkbook, Worksheet } from './types.js';
@@ -14,13 +15,15 @@ describe('worker render dependencies', () => {
     const math = {} as MathRenderer;
     const threeD = { render: vi.fn() } as unknown as ChartThreeDRenderer;
     const regionMap = { render: vi.fn() } as unknown as ChartRegionMapRenderer;
+    const chartEx = { render: vi.fn() } as unknown as ChartExRenderer;
 
-    expect(workerRenderDeps(ws, styles, { math, threeD, regionMap })).toEqual({
+    expect(workerRenderDeps(ws, styles, { math, threeD, regionMap, chartEx })).toEqual({
       ws,
       styles,
       math,
       threeD,
       regionMap,
+      chartEx,
     });
   });
 });
